dtfs-central-api/src/v1/controllers/utilisation-report-service/get-utilisation-reports.controller/helpers.ts
import { isValidReportPeriod } from '../../../validation/utilisation-report-service/utilisation-report-validator';
import { ReportPeriod } from '../../../../types/utilisation-reports';

export const parseReportPeriod = (reportPeriod: string | undefined): ReportPeriod | undefined => {
  if (!reportPeriod) {
    return undefined;
  }

  const parsedReportPeriod = JSON.parse(reportPeriod) as unknown;
  if (isValidReportPeriod(parsedReportPeriod)) {
    return parsedReportPeriod;
  }
  return undefined;
};
